+++ src/routes/policy/openvpn.js
@@ -27,7 +27,7 @@ var router = express.Router();
 import { PolicyRuleToOpenVPN } from '../../models/policy/PolicyRuleToOpenVPN';
 import { Firewall } from '../../models/firewall/Firewall';
 import { PolicyCompilation } from '../../models/policy/PolicyCompilation';
-const policy_r__ipobjModel = require('../../models/policy/policy_r__ipobj');
+import { PolicyRuleToIPObj } from '../../models/policy/PolicyRuleToIPObj';
 import { PolicyRule } from '../../models/policy/PolicyRule';
 const fwcError = require('../../utils/error_table');
 const utilsModel = require("../../utils/utils.js");
@@ -66,7 +66,7 @@ async (req, res) => {
 			throw fwcError.ALREADY_EXISTS;
 
 		// Get content of positions.
-		const content = await policy_r__ipobjModel.getPositionsContent(req.dbCon, req.body.position, req.body.new_position);
+		const content = await PolicyRuleToIPObj.getPositionsContent(req.dbCon, req.body.position, req.body.new_position);
 		if (content.content1!=='O' || content.content2!=='O')
 			throw fwcError.BAD_POSITION;
